added moon. refactored Scene
